Add a basic command line interface to the client
The CLI supports simple operations like:

    ./cli.py forms list
    ./cli.py forms list --author=someone

Configuration values for the client can be specified via command line
arguments or via a per-user configuration file for ease of use.

The CLI does not use argparse, but that is by design so that we can
dynamically lookup which API method to call via the CLI instead of
hard-coding the values. This minimizes the amount of maintenance of the
CLI going forward as we don't have to update it whenever the client gets
changed.

This is a first step towards implementing #18